requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
environ==1.0
greenlet==1.1.0
idna==3.2
mysqlclient==2.0.3
numpy==1.21.1
pandas==1.3.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
sqlparse==0.4.1
urllib3==1.26.6
xmltodict==0.12.0
